fix: refactor utils
