+++ src/index.ts
@@ -2,22 +2,7 @@ import path from 'path';
 import fs from 'fs-extra';
 import recursiveReaddir from 'recursive-readdir';
 import { LoadContext, Props } from '@docusaurus/types';
-
-const absoluteUrlRegExp = /(href|src)="(?!http[s]|ftp?:\/\/)([^"|#]+)"/g;
-const isDirectory = (dirPath: string) => path.extname(dirPath) === '';
-
-const convertAbsolutePathsToRelative = (content: string, filePath: string) =>
-  content.replace(absoluteUrlRegExp, (_absoluteUrl: string, $1: string, $2: string) => {
-    const currentDirPath = path.dirname(filePath);
-    const relativeDirPath = currentDirPath === '.' ? '.' : path.relative(currentDirPath, '');
-
-    let relativePath = path.join(relativeDirPath, $2);
-    if (isDirectory(relativePath)) {
-      relativePath = path.join(relativePath, 'index.html');
-    }
-
-    return `${$1}='${relativePath}'`;
-  });
+import { convertAbsolutePathsToRelative } from './utils/convertPath';
 
 const websiteTextualFileExtensions = ['.css', '.js', '.html', '.xml'];
 const isNotWebsiteTextualFile = (filePath: string, stats: any) =>
@@ -35,6 +20,7 @@ export default (_context: LoadContext, _options: any) => {
           await fs.writeFile(filePath, convertAbsolutePathsToRelative(String(content), relativePath));
         })
       );
+      console.log('Success!: Converted absolute paths to relative.')
     },
   };
 }
